+++ backend/src/app.module.ts
@@ -25,6 +25,7 @@ import { MessageModule } from "./message/message.module";
 import { MessageModel } from "./message/models/message.model";
 import { ConversationModel } from "./message/models/conversation.model";
 import { ActiveConversationModel } from "./message/models/active_conversation.model";
+import { UserConversationModel } from "./message/models/users_conversation.model";
 
 @Module({
   providers: [
@@ -60,6 +61,7 @@ import { ActiveConversationModel } from "./message/models/active_conversation.mo
         MessageModel,
         ConversationModel,
         ActiveConversationModel,
+        UserConversationModel,
       ],
       autoLoadModels: true,
       synchronize: true,