+++ backend/src/message/message.module.ts
@@ -5,12 +5,18 @@ import { SequelizeModule } from "@nestjs/sequelize";
 import { MessageModel } from "./models/message.model";
 import { ActiveConversationModel } from "./models/active_conversation.model";
 import { ConversationModel } from "./models/conversation.model";
-import { MessageGateway } from './message.gateway';
+import { MessageGateway } from "./message.gateway";
+import { UserConversationModel } from "./models/users_conversation.model";
 
 @Module({
   imports: [
     AuthModule,
-    SequelizeModule.forFeature([MessageModel, ConversationModel, ActiveConversationModel]),
+    SequelizeModule.forFeature([
+      MessageModel,
+      ConversationModel,
+      ActiveConversationModel,
+      UserConversationModel,
+    ]),
   ],
   providers: [ConversationService, MessageGateway],
   exports: [ConversationService],